requirements.txt
aiogram~=2.20
asyncio
python_dotenv~=0.20
